requirements.txt
# Flask and core dependencies
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.11.0
Flask-Session==0.5.0

# Strands agents dependencies for MCP support
strands-agents-builder>=0.1.0
strands-agents>=1.0.0

# AWS SDK for Bedrock
boto3>=1.34.0
botocore>=1.34.0

# MCP (Model Context Protocol) dependencies
mcp>=1.0.0
anyio>=4.0.0

# Async and WebSocket support
eventlet==0.33.3

# Utilities
python-dotenv==1.0.0

# Optional - for better logging
rich>=14.0.0